move REAsset components into a separate folder
