+++ client/src/components/REAsset/AddREAssetForm.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect, useState } from 'react';
 import {
-  Checkbox,
   Button,
   Dialog,
   DialogHeader,
