+++ client/src/components/REAsset/REListItem.tsx
